cuda_optimized_tts.py
#!/usr/bin/env python3
"""
CUDA优化的TTS处理器
"""
import torch
import time
from typing import List

class CUDAGraphVocoder:
    """CUDA Graph优化的Vocoder"""
    
    def __init__(self, vocoder_model, device="cuda"):
        self.vocoder_model = vocoder_model
        self.device = device
        self.graph = None
        self.static_input = None
        self.static_output = None
        self.is_initialized = False
        self.input_shape = None
        
    def initialize_graph(self, input_shape: tuple, num_warmup: int = 5):
        """初始化CUDA Graph"""
        if self.is_initialized and self.input_shape == input_shape:
            return True
            
        print(f"🔄 初始化CUDA Graph (输入形状: {input_shape})...")
        
        try:
            # 预热GPU
            for _ in range(num_warmup):
                dummy_input = torch.randn(input_shape, device=self.device, dtype=torch.float16)
                with torch.no_grad():
                    _ = self.vocoder_model(dummy_input)
            
            torch.cuda.synchronize()
            
            # 创建静态输入输出
            self.static_input = torch.randn(input_shape, device=self.device, dtype=torch.float16)
            
            # 捕获CUDA Graph
            self.graph = torch.cuda.CUDAGraph()
            with torch.cuda.graph(self.graph):
                with torch.no_grad():
                    self.static_output = self.vocoder_model(self.static_input)
            
            self.is_initialized = True
            self.input_shape = input_shape
            print(f"✅ CUDA Graph初始化完成")
            return True
            
        except Exception as e:
            print(f"❌ CUDA Graph初始化失败: {e}")
            self.is_initialized = False
            return False
    
    def inference(self, input_tensor: torch.Tensor) -> torch.Tensor:
        """使用CUDA Graph进行推理"""
        input_shape = tuple(input_tensor.shape)
        
        # 检查是否需要重新初始化
        if not self.is_initialized or self.input_shape != input_shape:
            if not self.initialize_graph(input_shape):
                # 回退到普通推理
                with torch.no_grad():
                    return self.vocoder_model(input_tensor)
        
        # 复制输入到静态输入
        self.static_input.copy_(input_tensor)
        
        # 执行Graph
        self.graph.replay()
        
        return self.static_output.clone()

class OptimizedVocoderProcessor:
    """优化的Vocoder处理器"""
    
    def __init__(self, vocoder_model, device="cuda", chunk_size=500):
        self.vocoder_model = vocoder_model
        self.device = device
        self.chunk_size = chunk_size
        self.cuda_graph_vocoder = CUDAGraphVocoder(vocoder_model, device)
    
    def split_into_chunks(self, pred_spec: torch.Tensor) -> List[torch.Tensor]:
        """将频谱分割为块"""
        if pred_spec.shape[-1] <= self.chunk_size:
            return [pred_spec]
        
        chunks = []
        total_length = pred_spec.shape[-1]
        
        for start in range(0, total_length, self.chunk_size):
            end = min(start + self.chunk_size, total_length)
            chunk = pred_spec[:, :, start:end]
            chunks.append(chunk)
        
        return chunks
    
    def process_optimized(self, pred_spec: torch.Tensor) -> torch.Tensor:
        """优化处理主函数"""
        start_time = time.perf_counter()
        
        # 分割为块
        chunks = self.split_into_chunks(pred_spec)
        
        if len(chunks) == 1:
            # 单个块，使用CUDA Graph
            result = self.cuda_graph_vocoder.inference(chunks[0])
        else:
            # 多个块，逐个处理并合并
            audio_chunks = []
            for chunk in chunks:
                audio_chunk = self.cuda_graph_vocoder.inference(chunk)
                audio_chunks.append(audio_chunk)
            
            # 合并块
            result = torch.cat(audio_chunks, dim=-1)
        
        end_time = time.perf_counter()
        print(f"⚡ 优化处理耗时: {end_time - start_time:.3f}s")
        
        return result

if __name__ == "__main__":
    print("🚀 CUDA优化TTS处理器")
    print("主要优化:")
    print("1. CUDA Graph - 减少内核启动开销")
    print("2. 智能分块 - 处理长音频序列")
    print("3. 内存优化 - 减少GPU内存分配")
    print("\n预期性能提升: 2-5x")